Add termination proof of program_sem and update Hoare triple.
This change removes the "fuel" hack from program_sem to
simplify termination proof.
The termination argument introduced in this commit is
much more involved, as it requires showing that each
instruction executed either strictly decreases the
gas or is the last instruction of the execution.
The benefit is to have a stronger definition
of Hoare triple and avoiding having an existential
quantifier that is unwieldy to deal with in the proofs.
The previous Hoare triple definition was weaker, leading
to really odd statements being true, such as {P} c {P}
for any command "c" and any predicate "P".

This change breaks any proof that deals with program_sem,
including all lemmas that talk about triple.
I suspect that fixing the proofs should be fairly easy.