Fix Swap and Dup inst notation to match yellow paper.
Swap and Dup took an argument that did not match
the yellow paper notation (e.g. SWAP1 corresponds
to "Swap 0" in lem, etc.).
This can be confusing for someone unfamiliar with
the semantics and expecting one-to-one matching with
the yellow paper.

This commit investigates if we can get a more faithful
correspondence between the lem and yellow paper
formalisms.
It fixes the semantics to use the more natural notation
that follows the same convention as the yellow paper
and fixes the ocaml parser used to validate the
lem semantics.

All the previously successful tests pass
(tested running ./runVmTest.native 89).
However, there is more work required to fix
the Hoare triples for Swap and Dup, and
existing proofs, as well as fixing
the Isabelle bytecode parser to generate
the appropriate instruction constructors.